Use CMake targets and portable integer logging in autoware_map_loader
This upstreams RoboStack downstream patch `patch/ros-rolling-autoware-map-loader.osx.patch`.

Best-guess rationale: linking to yaml-cpp and fmt through imported targets is more robust across package managers, and portable PRIu64 logging avoids uint64_t format mismatches on platforms where unsigned long and unsigned long long differ.

+++ map/autoware_map_loader/src/lanelet2_map_loader/lanelet2_map_loader_node.cpp
@@ -38,6 +38,8 @@
 #include "lanelet2_map_loader_utils.hpp"
 #include "lanelet2_selected_map_loader_module.hpp"
 
+#include <cinttypes>
+
 #include <autoware/geography_utils/lanelet2_projector.hpp>
 #include <autoware/lanelet2_utils/conversion.hpp>
 #include <autoware_lanelet2_extension/io/autoware_osm_parser.hpp>
@@ -172,8 +174,8 @@ void Lanelet2MapLoaderNode::on_map_projector_info(
     if (map_major_ver > static_cast<uint64_t>(lanelet::autoware::version)) {
       RCLCPP_WARN(
         get_logger(),
-        "format_version(%ld) of the provided map(%s) is larger than the supported version(%ld)",
-        map_major_ver, lanelet2_filename.c_str(),
+        "format_version(%" PRIu64 ") of the provided map(%s) is larger than the supported version(%" PRIu64 ")",
+        static_cast<uint64_t>(map_major_ver), lanelet2_filename.c_str(),
         static_cast<uint64_t>(lanelet::autoware::version));
       if (!allow_unsupported_version) {
         throw std::invalid_argument(
